+++ src/org/sosy_lab/java_smt/basicimpl/AbstractStringFormulaManager.java
@@ -206,6 +206,13 @@ public abstract class AbstractStringFormulaManager<TFormulaInfo, TType, TEnv, TF
   protected abstract TFormulaInfo allImpl();
 
   @Override
+  public RegexFormula allChar() {
+    return wrapRegex(allCharImpl());
+  }
+
+  protected abstract TFormulaInfo allCharImpl();
+
+  @Override
   public RegexFormula range(StringFormula start, StringFormula end) {
     return wrapRegex(range(extractInfo(start), extractInfo(end)));
   }
@@ -235,7 +242,7 @@ public abstract class AbstractStringFormulaManager<TFormulaInfo, TType, TEnv, TF
 
   @Override
   public RegexFormula intersection(RegexFormula regex1, RegexFormula regex2) {
-    return wrapRegex(union(extractInfo(regex1), extractInfo(regex2)));
+    return wrapRegex(intersection(extractInfo(regex1), extractInfo(regex2)));
   }
 
   protected abstract TFormulaInfo intersection(TFormulaInfo pParam1, TFormulaInfo pParam2);
@@ -260,7 +267,7 @@ public abstract class AbstractStringFormulaManager<TFormulaInfo, TType, TEnv, TF
   }
 
   protected TFormulaInfo difference(TFormulaInfo pParam1, TFormulaInfo pParam2) {
-    return union(pParam1, complement(pParam2));
+    return intersection(pParam1, complement(pParam2));
   }
 
   @Override